♻️ stock 분리
