+++ src/basic/features/stock-status/lib.js
@@ -1,0 +1,4 @@
+import { isOutOfStock } from '../../entities/stock/model.js';
+
+export const getStockStatusMessage = (item) =>
+  isOutOfStock(item) ? '품절' : `재고 부족 (${item.quantity}개 남음)`;
