+++ src/basic/features/stock-status/ui.js
@@ -1,0 +1,8 @@
+import { getStockStatusMessage } from './lib.js';
+import { isLowStock } from '../../entities/stock/model.js';
+
+const formatItemStockDisplay = (item) =>
+  `${item.name}: ${getStockStatusMessage(item)}\n`;
+
+export const updateStockInformation = (prodList) =>
+  prodList.filter(isLowStock).map(formatItemStockDisplay).join('');
