+++ src/basic/shared/lib/stock/config.js
@@ -1,0 +1,4 @@
+export const STOCK = Object.freeze({
+  LOW: 5,
+  EMPTY: 0,
+});
